Update D. Difference.cpp

+++ C++ Assignments/Codeforces/Sheet1/D. Difference.cpp
@@ -4,7 +4,7 @@ using namespace std;
 
 int main()
 {
-  int A,B,C,D,X;
+  long long int A,B,C,D,X;
   cin>>A>>B>>C>>D;
   X = (A*B) - (C*D);
   cout<<"Difference = "<<X;
